拆分公开和非公开代码，将errors.New方法全部换成fmt.Errorf，提高将来可能的灵活性
